feat: divided the task into separate files
