+++ app/pug/index.pug
@@ -1,7 +1,7 @@
 extends templates/layout
 
 block mainContent
-  h2 Welcome to the Gulpify.
+  h2 Welcome to the Gulpify. 
   p It is a simple boilerplate to automate your front-end workflow.
   ul
     each item in advantages
