+++ app/pug/templates/base.pug
@@ -4,7 +4,7 @@ html(lang="en")
     meta(charset="UTF-8")
     meta(name="viewport", content="width=device-width, initial-scale=1.0")
     title Gulpify
-    link(rel="stylesheet", href="/styles.css")
+    link(rel="stylesheet", href="/css/main.css")
   body
     include ../components/header.pug
     block content
