+++ app/scss/main.scss
@@ -1,0 +1,8 @@
+h2 {
+  color: red;
+}
+
+li {
+  color: blue;
+  font-size: 30px;
+}
