+++ tasks/buildSCSS.js
@@ -1,0 +1,32 @@
+import { src, dest } from 'gulp';
+import gulpSass from 'gulp-sass';
+import sass from 'sass';
+import autoprefixer from 'gulp-autoprefixer';
+import cleanCSS from 'gulp-clean-css';
+import sourcemaps from 'gulp-sourcemaps';
+import plumber from 'gulp-plumber';
+import notify from 'gulp-notify';
+import browserSync from 'browser-sync';
+
+const server = browserSync.create();
+const sassCompiler = gulpSass(sass);
+
+const buildSCSS = () =>
+  src('app/scss/**/*.scss')
+    .pipe(
+      plumber({
+        errorHandler: notify.onError((error) => ({
+          title: 'SCSS',
+          message: '<%= error.message %>',
+        })),
+      }),
+    )
+    .pipe(sourcemaps.init())
+    .pipe(sassCompiler().on('error', sassCompiler.logError))
+    .pipe(autoprefixer())
+    .pipe(cleanCSS())
+    .pipe(sourcemaps.write('.'))
+    .pipe(dest('build/css'))
+    .pipe(server.stream());
+
+export default buildSCSS;
