+++ tasks/clear.js
@@ -1,0 +1,5 @@
+import { deleteAsync } from 'del';
+
+const clear = () => deleteAsync('./build');
+
+export default clear;
